Feature: Move type to shared directory
The app did not build correctly when the shared type was in the ts-backend folder
due to needing to process the entire contents of that file, so instead we move
the type to a shared location.

In the future we may define in one spot and then move automatically as a prepare step, but
this should be okay for now.
