+++ ts-backend/src/index.ts
@@ -12,6 +12,7 @@ import {
   addRepositoriesToResult,
 } from "./fetchers";
 import { checkRateLimit, CustomOctokit, personalOctokit } from "./lib/octokit";
+import { RepositoryResult } from '../../types'
 
 export interface Result {
   meta: {
@@ -29,43 +30,6 @@ export interface Result {
     // "teamsCount": number;
   };
   repositories: Record<string, RepositoryResult>;
-}
-
-export interface RepositoryResult {
-  // Repo metadata
-  repositoryName: string;
-  repoNameWithOwner: string;
-  licenseName: string;
-  topics: string[];
-
-  // Counts of various things
-  projectsCount: number;
-  projectsV2Count: number;
-  discussionsCount: number;
-  forksCount: number;
-  totalIssuesCount: number;
-  openIssuesCount: number;
-  closedIssuesCount: number;
-  totalPullRequestsCount: number;
-  openPullRequestsCount: number;
-  closedPullRequestsCount: number;
-  mergedPullRequestsCount: number;
-  watchersCount: number;
-  starsCount: number;
-  collaboratorsCount: number;
-
-  // Flags
-  discussionsEnabled: boolean;
-  projectsEnabled: boolean;
-  issuesEnabled: boolean;
-
-  // Calculated metrics
-  openIssuesAverageAge: number;
-  openIssuesMedianAge: number;
-  closedIssuesAverageAge: number;
-  closedIssuesMedianAge: number;
-  issuesResponseAverageAge: number;
-  issuesResponseMedianAge: number;
 }
 
 export type Fetcher = (
